refactor(services): разделён refresh pipeline лент
FeedRefreshService оставлен контейнером зависимостей и публичного FeedRefreshCoordinating contract
conditional request, runtime refresh flow, not-modified handling, fetched response reconciliation, diagnostics log persistence и batch orchestration вынесены в отдельные файлы
поведение refresh flow сохранено, но границы сетевого ответа, persistence updates и batch execution стали явнее
